iOS: compile fixes for the file-backed open (#147)
Int32 for the core's status in a closure, and FPDF_ERR_FILE as its value in
the app and tests, which do not import CPdfium. Verified on the Mac mini:
iOS simulator build and all 90 tests pass, including the new engine tests
and Save a copy from a staged file.

+++ ios/MegaPDF/Engine/PdfEngine.swift
@@ -153,7 +153,7 @@ actor PdfEngine {
         guard !document.isDestroyed else { throw PdfError.saveFailed }
         let copy = FileManager.default.temporaryDirectory.appendingPathComponent("reading-\(UUID().uuidString).pdf")
         let status = copy.withUnsafeFileSystemRepresentation { path -> Int32 in
-            guard let path else { return MEGAPDF_ERR_ARGUMENT }
+            guard let path else { return Int32(MEGAPDF_ERR_ARGUMENT) }
             return megapdf_read_from_copy(document.core, path)
         }
         guard status == MEGAPDF_OK else { throw PdfError.saveFailed }
